# Improve the performance of stats computation by reducing branches