--- a/layout/chunk.go
+++ b/layout/chunk.go
@@ -24,6 +24,7 @@ func PagesToChunk(pages []*Page) *Chunk {
 	var maxVal interface{} = pages[0].MaxVal
 	var minVal interface{} = pages[0].MinVal
 	pT, cT := pages[0].Schema.Type, pages[0].Schema.ConvertedType
+	funcTable := common.FindFuncTable(pT, cT)
 
 	for i := 0; i < ln; i++ {
 		if pages[i].Header.DataPageHeader != nil {
@@ -33,8 +34,8 @@ func PagesToChunk(pages []*Page) *Chunk {
 		}
 		totalUncompressedSize += int64(pages[i].Header.UncompressedPageSize) + int64(len(pages[i].RawData)) - int64(pages[i].Header.CompressedPageSize)
 		totalCompressedSize += int64(len(pages[i].RawData))
-		maxVal = common.Max(maxVal, pages[i].MaxVal, pT, cT)
-		minVal = common.Min(minVal, pages[i].MinVal, pT, cT)
+		minVal = common.Min(funcTable, minVal, pages[i].MinVal)
+		maxVal = common.Max(funcTable, maxVal, pages[i].MaxVal)
 	}
 
 	chunk := new(Chunk)
@@ -81,6 +82,7 @@ func PagesToDictChunk(pages []*Page) *Chunk {
 	var maxVal interface{} = pages[1].MaxVal
 	var minVal interface{} = pages[1].MinVal
 	pT, cT := pages[1].Schema.Type, pages[1].Schema.ConvertedType
+	funcTable := common.FindFuncTable(pT, cT)
 
 	for i := 0; i < len(pages); i++ {
 		if pages[i].Header.DataPageHeader != nil {
@@ -91,8 +93,8 @@ func PagesToDictChunk(pages []*Page) *Chunk {
 		totalUncompressedSize += int64(pages[i].Header.UncompressedPageSize) + int64(len(pages[i].RawData)) - int64(pages[i].Header.CompressedPageSize)
 		totalCompressedSize += int64(len(pages[i].RawData))
 		if i > 0 {
-			maxVal = common.Max(maxVal, pages[i].MaxVal, pT, cT)
-			minVal = common.Min(minVal, pages[i].MinVal, pT, cT)
+			minVal = common.Min(funcTable, minVal, pages[i].MinVal)
+			maxVal = common.Max(funcTable, maxVal, pages[i].MaxVal)
 		}
 	}