Now instead of checking types every time, we check types exactly once. Also removes reflection checking for sizeof.

This is just shy of a 15% win on the benchmark in #257:

```
name         old time/op    new time/op    delta
WriteCSV-44    10.2ms ± 7%     8.7ms ± 3%  -14.73%  (p=0.000 n=19+19)

name         old alloc/op   new alloc/op   delta
WriteCSV-44    10.7MB ± 0%    10.7MB ± 0%   -0.00%  (p=0.025 n=20+18)

name         old allocs/op  new allocs/op  delta
WriteCSV-44      103k ± 0%      103k ± 0%     ~     (p=0.523 n=20+20)
```