@alex 2020-05-10
This involved significant changes to `common.go` -- how well is that file covered by tests? I'm slightly concerned that I implemented this correctly for some of the more obscure types (int96, interval, and decimal).