@alex 2020-05-12
With the other performance improvements this has a proportionally greater impact:

```
name         old time/op    new time/op    delta
WriteCSV-44    6.19ms ± 2%    4.83ms ± 2%  -21.92%  (p=0.000 n=20+19)

name         old alloc/op   new alloc/op   delta
WriteCSV-44    5.44MB ± 0%    5.44MB ± 0%     ~     (p=0.658 n=19+18)

name         old allocs/op  new allocs/op  delta
WriteCSV-44     61.5k ± 0%     61.5k ± 0%     ~     (all equal)
```