@alex 2020-05-12
I'm not sure how you think about this PR, but I'd recommend reviewing/merging #261 first so I can measure this on top of that.