@alex 2020-05-13
Sure thing, this PR is definitely the most complicated! If you can think of ways to simplify it, please let me know. Updated performance numbers:

```
name         old time/op    new time/op    delta
WriteCSV-44    5.96ms ±11%    5.38ms ±17%  -9.86%  (p=0.004 n=18+20)

name         old alloc/op   new alloc/op   delta
WriteCSV-44    5.44MB ± 0%    5.44MB ± 0%    ~     (p=0.381 n=18+20)

name         old allocs/op  new allocs/op  delta
WriteCSV-44     61.5k ± 0%     61.5k ± 0%    ~     (all equal)
```