@xitongsys 2020-05-13
> I removed a few of the tests for `nil` handling -- this is because these methods no longer handle `nil`. I believe it's not possible for a `nil` interface to actually be passed in any of the places that use `Min`/`Max`/`MinMaxSize`.

hi, @alex 
The `nil` should be handled. When we calculate the min/max value of `table.Values`, the element can be `nil`.