@alex 2020-05-13
Ok! I have an idae about how to implement this, and probably reduce the amount of code. I think it will cost a few % of performance, but hopefully still a decent win.