@alex 2020-05-13
Ok, it should now handle `nil` correctly in `Min`/`Max`. It gave back a bit of perf, but the code is much cleaner:

```
name         old time/op    new time/op    delta
WriteCSV-44    5.20ms ± 1%    4.93ms ± 2%  -5.11%  (p=0.000 n=18+20)

name         old alloc/op   new alloc/op   delta
WriteCSV-44    4.76MB ± 0%    4.76MB ± 0%    ~     (p=0.744 n=20+19)

name         old allocs/op  new allocs/op  delta
WriteCSV-44     61.5k ± 0%     61.5k ± 0%    ~     (all equal)
```